Tidied up BATT_LOW shutdown script, and added shutdown delay Signal now needs to be low continuously for two seconds before shutdown is triggered

+++ batt_low.py
@@ -4,12 +4,25 @@ import os
 import time
 import RPi.GPIO as GPIO
 
+# Change this number if using a different GPIO pin
 pin = 27
 
-GPIO.setmode(GPIO.BCM)   # Use Broadcom (BCM) pin numbering
-GPIO.setup(pin, GPIO.IN) # Input : listen for pin
+# Set up switch pin for input
+GPIO.setmode(GPIO.BCM)
+GPIO.setup(pin, GPIO.IN)
 
-GPIO.wait_for_edge(pin, GPIO.FALLING)
-print("BATT_LOW shutdown request received")
-os.system("shutdown -h now")   # Software system shutdown (commence immediately)
-
+while True:
+	# Wait indefinitely for a falling edge
+	GPIO.wait_for_edge(pin, GPIO.FALLING)
+	print("BATT_LOW shutdown request received")
+	
+	# Wait for up to 2 seconds for a rising edge (timeout is in milliseconds)
+	result = GPIO.wait_for_edge(pin, GPIO.RISING, timeout=2000)
+	
+	if result is None and not GPIO.input(pin):
+		print('BATT_LOW still asserted, time to shutdown...')
+		os.system("shutdown -h now")
+	else:
+		print('BATT_LOW signal changed, shutdown cancelled.')
+	
+	time.sleep(0.1)
